gpu: nvgpu: Cast negative int to u32 before shift
A shift of a negative number is undefined; so to work around
said warning simply cast to a u32 first. In this case the
resulting operation should be ok since the sign bits are
maintained when the 32 bit negative integer is shifted into
a 24 bit negative integer.

JIRA NVGPU-525

Reviewed-on: https://git-master.nvidia.com/r/1673826
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/gp106/therm_gp106.c
@@ -35,7 +35,7 @@
 void gp106_get_internal_sensor_limits(s32 *max_24_8, s32 *min_24_8)
 {
 	*max_24_8 = (0x87 << 8);
-	*min_24_8 = ((-216) << 8);
+	*min_24_8 = (((u32)-216) << 8);
 }
 
 int gp106_get_internal_sensor_curr_temp(struct gk20a *g, u32 *temp_f24_8)
